refactor: remove subcomponents folder from header folder
